# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Demo Gin Server API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/luxuze/demo-gin-server",
            "email": "luxuze1994@gmial.com"
        },
        "license": {
            "name": "DO WHAT THE FUCK YOU WANT TO PUBLIC LICENSE",
            "url": "https://github.com/luxuze/demo-gin-server/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/login/register": {
            "post": {
                "description": "add user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "user",
                "parameters": [
                    {
                        "description": "add user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/db.TBL_USERS"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.TBL_USERS"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/defs.HTTPError"
                        }
                    }
                }
            }
        },
        "/wx": {
            "get": {
                "description": "Request WX get resp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WX"
                ],
                "summary": "Wx",
                "parameters": [
                    {
                        "type": "string",
                        "default": "123",
                        "description": "Request WX",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.TBL_USERS": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "may"
                },
                "phone": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string",
                    "example": "123"
                }
            }
        },
        "defs.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    }
}